Add Pause Menu to Game
In the pause menu we can access to options, resume game or return to
main menu.
Would be cool to add a fade animation when going in to the menu.
Maybe also change the music to something less loud?
Add button changes to options menu
Add that the music slider has the current value
Add an accesibility menu to make it easier